move metric functions to metrics
